cleanup custom type definitions + remove extended tsconfig.json
