+++ frontend/app/components/host/CreateMatchForm.tsx
@@ -22,6 +22,8 @@ import { HostingRules } from '../HostingRules';
 import { PotentialConflicts } from './PotentialConflicts';
 import { If } from '../If';
 import { SwitchField } from '../fields/SwitchField';
+import { ReactDatePickerProps } from 'react-datepicker';
+import { RcTimePickerProps } from 'rc-time-picker';
 
 const noop = (): void => undefined;
