--- frontend/app/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "compilerOptions": {
-    "noImplicitReturns": true,
-    "noImplicitAny": true,
-    "noFallthroughCasesInSwitch": true
-  },
-  "extends": "../tsconfig.json",
-  "include": [
-    "./**/*.tsx", "./**/*.ts"
-  ]
-}
