+++ frontend/typings/snuownd.d.ts
@@ -1,0 +1,12 @@
+declare module 'snuownd' {
+  type SnuOwndParser = {
+    readonly render(markdown: string): string;
+  };
+
+  type SnuOwnd = {
+    readonly getParser(): SnuOwndParser;
+  };
+
+  const value: SnuOwnd;
+  export = value;
+}
